==> sources/05-ds-taxonomy-preproc.ipynb <==
from lib import *

---


tax_raw = (
    pd.read_excel('acm_css_taxonomy_modified.xlsx', header=None)
     .rename(columns={0:'level'}).iloc[3:459]
)

tax_raw.shape

---

topics = set(tax_raw.iloc[:, 1:7].fillna('').sum(axis=1).str.strip().str.lower())

len(topics)

---

ds_tax = (
    tax_raw
    .fillna('')
    .assign(
        label=lambda x: x.iloc[:, 1:7].sum(axis=1).str.strip(),
        new=lambda x: x['level'].str.contains('\*'),
        level=lambda x: x['level'].str.replace(r'*', ''),
    )
    .assign(
        lvls=lambda x: x['level'].str.split('.')\
                                  .apply(lambda x: [y for y in x if y.isdecimal()]))
    .assign(
        depth=lambda x: x['lvls'].apply(len)
    )
    .loc[lambda x: x['label'] != '', ['level', 'label', 'depth', 'lvls', 'new']]
)

ds_tax.head()

root = anytree.Node('root')

nodes = [root]
prev_depth = 0
parent = root
for i, s in ds_tax.iterrows():
    curr_depth = s['depth']
    if curr_depth < prev_depth:
        
        for i in range(prev_depth - curr_depth):
            parent = parent.parent
            
    elif curr_depth > prev_depth:
        if curr_depth - prev_depth > 1:
            print(s)
            raise ValueError
        parent = nodes[-1]
        
    n = anytree.Node(s['level'] + ' -- ' + s['label'], raw_name=s['label'],parent=parent)
    nodes.append(n)
    
    prev_depth = curr_depth




# for pre, fill, node in anytree.RenderTree(root):
#     print(pre, node.name)


from anytree.exporter import JsonExporter
exporter = JsonExporter(indent=2)

with open('ds_taxonomy.json', 'w') as f:
    f.write(exporter.export(root))

leaves = [i.raw_name for i in anytree.LevelOrderIter(root, filter_=lambda x: x.is_leaf)]

len(leaves)

ds_tax['isleaf'] = ds_tax['label'].isin(leaves)

ds_tax = (ds_tax
    .assign(label=lambda x: (x['level'] + ' -- ' + x['label']))
         )

ds_tax.to_csv('data/ds_taxonomy_parsed_full.csv', index=False)

---

ds_tax.loc[lambda x: x.isleaf, 'new'].sum()

ds_tax.loc[:,  'new'].sum()

ds_tax.shape

---

from pipe import *

(ds_tax.loc[lambda x: x['depth'] <= 2, ['level', 'label', 'depth']]
.assign(
    level=lambda x: x['level'].str.replace(' ', '')\
                              .str.replace('\+0', ''),
    label=lambda x: (x['depth'].apply(lambda y: '~~~~'*y) + x['label'])
)
 .loc[:, ['level', 'label']]
 .to_latex(index=False, escape=False)

) | stdout

(ds_tax.loc[:,  ['level', 'label', 'depth']]
.assign(
    level=lambda x: x['level'].str.replace(' ', '')\
                              .str.replace('\+0', ''),
#     label=lambda x: (x['depth'].apply(lambda y: '~~~~'*y) + x['label'])
)
#  .loc[:, ['level', 'label']]
#  .to_latex(index=False, escape=False, longtable=True)
.set_index(['level', 'depth'], append=True)
.unstack()
.fillna('')
 .reset_index()
 .drop('level_0', axis=1)
 .to_latex(longtable=True, index=False)
 ) | stdout

---
Plot

from anytree import Resolver

node = Resolver().get(root, '/root/5.+0 -- Computing methodologies/5.1. -- Artificial intelligence/5.1.1. -- Natural language processing')

node = Resolver().get(root, '/root/5.+0 -- Computing methodologies/5.1. -- Artificial intelligence/5.1.3. -- Computer vision')

node.children

from anytree.exporter import DotExporter
dot = DotExporter(node)

dot.to_dotfile('tax_2.dot')

!dot -Tpdf tax_2.dot -o tax_2.pdf


  rankdir="LR";
    node [ fontname="arial bold", fontsize=20];

!dot -Tpdf tax.dot -o tax.pdf

---

root.depth

node.depth

from anytree import LevelOrderIter

for n in LevelOrderIter(root):
    if n.depth > 2:
        n.parent=None


dot = DotExporter(root)

dot.to_dotfile('tax_2upper.dot')

!dot -Tpdf tax_2upper.dot -o tax_2upper.pdf
