==> sources/07-text-to-topic-matrix.ipynb <==
%run lib.py

import build_ast
import text_util

# load text

ds_tax = pd.read_csv('data/ds_taxonomy_parsed_full.csv')
leaves = ds_tax.loc[ds_tax['isleaf'], 'label'].values

score_list = []
with shelve.open('asts_relevant_topic_scores') as scores:
    for k in scores:
        score_list.append(scores[k])

score_list_full = np.array(sum(score_list, []))
del score_list

indices = score_list_full[:, 0].astype(int)
score_array = np.vstack(score_list_full[:, 1])

score_df = pd.DataFrame(data=score_array, index=indices, columns=leaves)

score_df.head()

score_df.shape

score_df.isnull().sum().sum()

score_df.to_csv('data/text_to_topic_scores.csv')

---
