==> sources/09-thematic-clusters-lifting.ipynb <==
%run lib.py

---

from pipe import *
from taxonomy import print_tree
import pgfs
from anytree.importer import JsonImporter

importer =  JsonImporter()

with open('ds_taxonomy.json', 'r') as f:
    tax = importer.read(f)

print_tree(tax)

clusters = pd.read_excel('data/cluster_topics_u_nonzero.xlsx', index_col=[0,1])

cluster_memberships = (
    clusters.loc[5, 'u']
    .loc[lambda x: x > 0.1] # filter noise
    .pipe(lambda x: x / np.sqrt((x ** 2).sum())) # normalize
    .to_dict()
)
cluster_elements = set(list(cluster_memberships.keys()))

lifter = pgfs.PGFS(tax, lmbda=0.1, gamma=0.9)

lifted = lifter.lift(cluster_memberships)

cluster_elements

gaps = lifted.L | as_list | sort
heads = lifted.H - cluster_elements | as_list | sort 
offs = cluster_elements & lifted.H | as_list | sort 
clust =  [(i, round(j, 3)) for i, j in cluster_memberships.items()]

gaps_str = gaps | concat('\n')
heads_str = heads | concat('\n')
offs_str = offs | concat('\n')
clust_str =  clust | concat('\n') 


print(f'Cluster elements: \n{clust_str}\n\n'
      f'Head subjects: \n{heads_str}\n\n'
      f'Offshoots: \n{offs_str}\n\n'
      f'Gaps: \n{gaps_str}\n\n'
      f'p = {lifted.p}')
      

---

Batch

lifter = pgfs.PGFS(tax, lmbda=0.1, gamma=0.9)

lifting_df = []
lifted_trees = []

for cluster_id in clusters.index.levels[0]:
    
    cluster_memberships = (
        clusters.loc[cluster_id, 'u']
        .loc[lambda x: x > 0.1] # filter noise
        .pipe(lambda x: x / np.sqrt((x ** 2).sum())) # normalize
        .to_dict()
    )
    cluster_elements = set(list(cluster_memberships.keys()))

    lifted = lifter.lift(cluster_memberships)

    gaps = lifted.L | as_list | sort
    heads = lifted.H - cluster_elements | as_list | sort 
    offs = cluster_elements & lifted.H | as_list | sort 
    clust =  [(i, round(j, 3)) for i, j in cluster_memberships.items()]

    df = (
        pd.concat((
            pd.Series(clust, name='topics_and_memberships'), 
            pd.Series(heads, name='head_subjects'), 
            pd.Series(gaps, name='gaps'), 
            pd.Series(offs, name='offshoots')), axis=1)
        .assign(cluster_id=cluster_id)
        .set_index('cluster_id', append=True)
        .reorder_levels([1,0])
    )

    lifting_df.append(df)
    lifted_trees.append(lifted)

lifting_df = pd.concat(lifting_df)


lifting_df.to_excel('data/lifting_results.xlsx', index=True)


joblib.dump(lifted_trees, 'data/lifted_trees.pkl')
